backend/requirements.txt
# FastAPI y servidor
fastapi==0.109.0
uvicorn[standard]==0.27.0
python-multipart==0.0.6

# Base de datos
sqlalchemy==2.0.25
psycopg2-binary==2.9.9
alembic==1.13.1

# Autenticación y seguridad
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-dotenv==1.0.0
pydantic-settings==2.1.0

# Validación
pydantic==2.5.3
email-validator==2.1.0

# Generación de PDFs
reportlab==4.0.9
pypdf2==3.0.1

# Utilidades
python-dateutil==2.8.2
pytz==2024.1
